Clean up code, add getHumanReadableDate function
