+++ src/helpers/transactions.js
@@ -135,6 +135,23 @@ const buildTransactionsSections = (
   };
 };
 
+export function getHumanReadableDate(date) {
+  const todayTimestamp = calculateTimestampOfToday();
+  const yesterdayTimestamp = calculateTimestampOfYesterday();
+  const thisYearTimestamp = calculateTimestampOfThisYear();
+
+  const timestamp = new Date(date * 1000);
+
+  return format(
+    timestamp,
+    timestamp > todayTimestamp
+      ? `'Today'`
+      : timestamp > yesterdayTimestamp
+      ? `'Yesterday'`
+      : `'on' MMM d${timestamp > thisYearTimestamp ? '' : ' yyyy'}`
+  );
+}
+
 export const buildTransactionsSectionsSelector = createSelector(
   [
     contactsSelector,
